requirements.txt
fastapi~=0.115
uvicorn~=0.30
pydantic~=2.7
bandit~=1.8
safety~=3.2
langgraph~=0.1
langchain-core~=0.2
httpx~=0.27
pytest~=8.2
pylint~=3.2
mkdocs~=1.6
playwright
opencv-python-headless
easyocr
firebase-admin
python-dotenv
